Fix potential overflow and reduce memory in HandoverStorage
- Use static_cast<std::size_t> in span accessor pointer arithmetic
  to prevent potential integer overflow for large mn * ns products
- Only allocate lthreed-dependent arrays (rmnss_*, zmncs_*, lmncs_*,
  rmncs_*, zmnss_*, lmnss_*) when lthreed=true, saving memory for
  axisymmetric (2D) cases

+++ src/vmecpp/cpp/vmecpp/vmec/handover_storage/handover_storage.h
@@ -171,22 +171,28 @@ class HandoverStorage {
 
   // Span accessors for tri-diagonal solver (returns view of radial slice)
   inline std::span<double> TridiagAr(int mn) noexcept {
-    return std::span<double>(all_ar.data() + mn * ns_, ns_);
+    return std::span<double>(all_ar.data() + static_cast<std::size_t>(mn) * ns_,
+                            ns_);
   }
   inline std::span<double> TridiagAz(int mn) noexcept {
-    return std::span<double>(all_az.data() + mn * ns_, ns_);
+    return std::span<double>(all_az.data() + static_cast<std::size_t>(mn) * ns_,
+                            ns_);
   }
   inline std::span<double> TridiagDr(int mn) noexcept {
-    return std::span<double>(all_dr.data() + mn * ns_, ns_);
+    return std::span<double>(all_dr.data() + static_cast<std::size_t>(mn) * ns_,
+                            ns_);
   }
   inline std::span<double> TridiagDz(int mn) noexcept {
-    return std::span<double>(all_dz.data() + mn * ns_, ns_);
+    return std::span<double>(all_dz.data() + static_cast<std::size_t>(mn) * ns_,
+                            ns_);
   }
   inline std::span<double> TridiagBr(int mn) noexcept {
-    return std::span<double>(all_br.data() + mn * ns_, ns_);
+    return std::span<double>(all_br.data() + static_cast<std::size_t>(mn) * ns_,
+                            ns_);
   }
   inline std::span<double> TridiagBz(int mn) noexcept {
-    return std::span<double>(all_bz.data() + mn * ns_, ns_);
+    return std::span<double>(all_bz.data() + static_cast<std::size_t>(mn) * ns_,
+                            ns_);
   }
 
   // RHS span accessor: returns pointer to start of RHS block for mode mn
